+++ test/v1x15_to_v1x19/tasks_test.dart
@@ -36,6 +36,8 @@ Uri _uriUpdatePort(Uri uri, int port) {
 }
 
 void main() {
+  doCheckSurplusItems = true;
+
   DockerConnection connection;
   CreateResponse createdContainer;